+++ src/DifferentialModulator.cpp
@@ -52,11 +52,11 @@ int DifferentialModulator::process(std::vector<Buffer*> dataIn, Buffer* dataOut)
 {
 #ifdef DEBUG
     fprintf(stderr, "DifferentialModulator::process (dataIn:");
-    for (unsigned i = 0; i < dataIn.size(); ++i) {
+    for (size_t i = 0; i < dataIn.size(); ++i) {
         fprintf(stderr, " %p", dataIn[i]);
     }
     fprintf(stderr, ", sizeIn: ");
-    for (unsigned i = 0; i < dataIn.size(); ++i) {
+    for (size_t i = 0; i < dataIn.size(); ++i) {
         fprintf(stderr, " %zu", dataIn[i]->getLength());
     }
     fprintf(stderr, ", dataOut: %p, sizeOut: %zu)\n", dataOut, dataOut->getLength());
